refactoring: changed project structure
