+++ coursework_3/data_providers/data_provider.py
@@ -1,4 +1,4 @@
-from entities import Operation
+from coursework_3.entities import Operation
 
 
 class DataProvider:
