+++ coursework_3/entities/__init__.py
@@ -1,5 +1,5 @@
-from entities import operation
-from entities import operation_amount
+from coursework_3.entities import operation
+from coursework_3.entities import operation_amount
 
 Operation = operation.Operation
 OperationAmount = operation_amount.OperationAmount
