+++ coursework_3/entities/operation_amount.py
