+++ coursework_3/settings.py
@@ -1,0 +1,4 @@
+import os.path
+
+OPERATIONS_DATA_FILE = os.path.join("../data", "operations.json")
+SHOW_OPERATIONS_COUNT = 5
