+++ coursework_3/views/__init__.py
@@ -1,3 +1,3 @@
-from views import operation_view
+from coursework_3.views import operation_view
 
 OperationView = operation_view.OperationView
