--- settings.py
@@ -1,4 +1,0 @@
-import os.path
-
-OPERATIONS_DATA_FILE = os.path.join("data", "operations.json")
-SHOW_OPERATIONS_COUNT = 5
